+++ lib/legion/extensions/microsoft_teams.rb
@@ -48,6 +48,8 @@ if defined?(Legion::Extensions) &&
    Legion::Extensions.const_defined?(:Absorbers, false) &&
    Legion::Extensions::Absorbers.const_defined?(:Base, false)
   require_relative 'microsoft_teams/absorbers/meeting'
+  require_relative 'microsoft_teams/absorbers/chat'
+  require_relative 'microsoft_teams/absorbers/channel'
 end
 
 module Legion