+++ lib/legion/extensions/microsoft_teams/helpers/browser_auth.rb
@@ -106,14 +106,6 @@ module Legion
 
           private
 
-          def log
-            return Legion::Logging if defined?(Legion::Logging) # rubocop:disable Legion/HelperMigration/LoggingGuard
-
-            @log ||= Object.new.tap do |nl|
-              %i[debug info warn error fatal].each { |m| nl.define_singleton_method(m) { |*| nil } }
-            end
-          end
-
           def host_os
             RbConfig::CONFIG['host_os']
           end